requirements.txt
# 도로 이미지 기반 필지 추출 및 토지이용 계획 생성 프로젝트
# 필수 패키지 목록

# 이미지 처리
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# 시각화
matplotlib>=3.5.0

# LLM API
openai>=1.0.0



